Change intrinsics notation (breaking change)
Don't pass the output parameter as first argument, but return it instead
(as usual in python). This is a breaking change for all code that uses
intrinsics. You should get proper error messages though, because the
number of arguments changed and is recognized by pyright.

Examples:
before              after
pop(ra)             ra = pop()
sqrt(a,b)           a = sqrt(b)